# contracts/openapi-v1.yaml
openapi: 3.1.0
info:
  title: Sejong Civil AI API
  version: 4.0.0-draft
  description: |
    Final-scope draft. Validate through discovery/interview before implementation.
    Admin routes are local/private demo surfaces, not public authentication boundaries.
...
servers:
  - url: http://localhost:8000
paths:
...
  /api/v1/admin/failed-questions:
    get:
      operationId: listFailedQuestions
      parameters:
        - { $ref: '#/components/parameters/DemoActorId' }
        - { $ref: '#/components/parameters/DemoRole' }
        - { in: query, name: reason, schema: { $ref: '#/components/schemas/StoredFailureReason' } }
        - { in: query, name: status, schema: { $ref: '#/components/schemas/FailedQuestionStatus' } }
      responses:
        '200':
          description: Masked failures only
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FailedQuestionListResponse' }
        '403': { $ref: '#/components/responses/AdminError' }
  /api/v1/admin/failed-questions/{id}:
    get:
      operationId: getFailedQuestion
      parameters:
        - { $ref: '#/components/parameters/IdPath' }
        - { $ref: '#/components/parameters/DemoActorId' }
        - { $ref: '#/components/parameters/DemoRole' }
      responses:
        '200':
          description: Failed question detail
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FailedQuestionDetailResponse' }
        '403': { $ref: '#/components/responses/AdminError' }
        '404': { $ref: '#/components/responses/AdminError' }
  /api/v1/admin/failed-questions/{id}/reason:
    patch:
      operationId: confirmFallbackReason
      parameters:
        - { $ref: '#/components/parameters/IdPath' }
        - { $ref: '#/components/parameters/DemoActorId' }
        - { $ref: '#/components/parameters/DemoRole' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReasonConfirmationRequest' }
      responses:
        '200':
          description: Reason confirmed
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ReasonConfirmationResponse' }
        '403': { $ref: '#/components/responses/AdminError' }
        '404': { $ref: '#/components/responses/AdminError' }
        '409': { $ref: '#/components/responses/AdminError' }
        '422': { $ref: '#/components/responses/AdminError' }
  /api/v1/admin/civic-scope-gaps:
    get:
      operationId: listCivicScopeGaps
      parameters:
        - { $ref: '#/components/parameters/DemoActorId' }
        - { $ref: '#/components/parameters/DemoRole' }
        - { in: query, name: status, schema: { $ref: '#/components/schemas/CivicScopeGapStatus' } }
      responses:
        '200':
          description: Separate masked civic scope-gap review queue
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CivicScopeGapListResponse' }
        '403': { $ref: '#/components/responses/AdminError' }
        '422': { $ref: '#/components/responses/AdminError' }
  /api/v1/admin/civic-scope-gaps/{id}/review:
    patch:
      operationId: reviewCivicScopeGap
      parameters:
        - { $ref: '#/components/parameters/IdPath' }
        - { $ref: '#/components/parameters/DemoActorId' }
        - { $ref: '#/components/parameters/DemoRole' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CivicScopeGapReviewRequest' }
      responses:
        '200':
          description: Scope gap marked for planning or dismissal without creating KB
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CivicScopeGapReviewResponse' }
        '403': { $ref: '#/components/responses/AdminError' }
        '404': { $ref: '#/components/responses/AdminError' }
        '409': { $ref: '#/components/responses/AdminError' }
        '422': { $ref: '#/components/responses/AdminError' }
  /api/v1/admin/kb-candidates:
    get:
      operationId: listKBCandidates
      parameters: [ { $ref: '#/components/parameters/DemoActorId' }, { $ref: '#/components/parameters/DemoRole' } ]
      responses:
        '200':
          description: Candidate list
          content:
            application/json:
              schema: { $ref: '#/components/schemas/KBCandidateListResponse' }
        '403': { $ref: '#/components/responses/AdminError' }
    post:
      operationId: createKBCandidate
      parameters: [ { $ref: '#/components/parameters/DemoActorId' }, { $ref: '#/components/parameters/DemoRole' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/KBCandidateCreate' }
      responses:
        '201':
          description: Draft candidate created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/KBCandidateCreateResponse' }
        '403': { $ref: '#/components/responses/AdminError' }
        '409': { $ref: '#/components/responses/AdminError' }
        '422': { $ref: '#/components/responses/AdminError' }
  /api/v1/admin/kb-candidates/{id}/submit:
    post:
      operationId: submitKBCandidate
      parameters:
        - { $ref: '#/components/parameters/IdPath' }
        - { $ref: '#/components/parameters/DemoActorId' }
        - { $ref: '#/components/parameters/DemoRole' }
      responses:
        '200':
          description: Candidate pending approval
          content:
            application/json:
              schema: { $ref: '#/components/schemas/KBCandidateSubmitResponse' }
        '403': { $ref: '#/components/responses/AdminError' }
        '404': { $ref: '#/components/responses/AdminError' }
        '409': { $ref: '#/components/responses/AdminError' }
  /api/v1/admin/kb-candidates/{id}/review:
    patch:
      operationId: reviewKBCandidate
      parameters:
        - { $ref: '#/components/parameters/IdPath' }
        - { $ref: '#/components/parameters/DemoActorId' }
        - { $ref: '#/components/parameters/DemoRole' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CandidateReviewRequest' }
      responses:
        '200':
          description: Review applied; approval activates KB atomically
          content:
            application/json:
              schema: { $ref: '#/components/schemas/KBCandidateReviewResponse' }
        '403': { $ref: '#/components/responses/AdminError' }
        '404': { $ref: '#/components/responses/AdminError' }
        '409': { $ref: '#/components/responses/AdminError' }
        '422': { $ref: '#/components/responses/AdminError' }
...
components:
  parameters:
...
    IdPath:
      in: path
      name: id
      required: true
      schema: { type: string, format: uuid }
    DemoActorId:
      in: header
      name: X-Demo-Actor-Id
      required: true
      description: Local/private demo actor only; not an authentication credential.
      schema: { type: string, minLength: 1, maxLength: 100 }
    DemoRole:
      in: header
      name: X-Demo-Role
      required: true
      description: Local/private role switch only; reject when admin routes are not privately gated.
      schema: { type: string, enum: [OPERATOR, APPROVER] }
  responses:
    AdminError:
      description: Stable local/private admin error without request input echo.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/AdminErrorEnvelope' }
...
  schemas:
...
    SupportedIntent:
      type: string
      enum: [MOVE_IN_RESIDENT_REGISTRATION, CERTIFICATE_ISSUANCE, BULKY_WASTE, LOCAL_TAX_GENERAL]
...
    StoredFailureReason:
      type: string
      description: OUT_OF_SCOPE and CIVIC_SCOPE_GAP never create a failed_questions row.
      enum: [INSUFFICIENT_GROUNDING, PERSONAL_LOOKUP, LEGAL_JUDGMENT]
    FailedQuestionStatus:
      type: string
      enum: [NEW, REASON_CONFIRMED]
...
    AdminErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error: { $ref: '#/components/schemas/AdminError' }
    AdminError:
      oneOf:
        - { $ref: '#/components/schemas/AdminRouteDisabledError' }
        - { $ref: '#/components/schemas/AdminForbiddenError' }
        - { $ref: '#/components/schemas/AdminNotFoundError' }
        - { $ref: '#/components/schemas/AdminInvalidStateError' }
        - { $ref: '#/components/schemas/AdminValidationFailedError' }
    AdminRouteDisabledError:
      type: object
      additionalProperties: false
      required: [code, message, request_id, retryable]
      properties:
        code: { const: ADMIN_ROUTE_DISABLED }
        message: { const: 관리자 기능을 사용할 수 없습니다. }
        request_id: { type: string, format: uuid }
        retryable: { const: false }
    AdminForbiddenError:
      type: object
      additionalProperties: false
      required: [code, message, request_id, retryable]
      properties:
        code: { const: ADMIN_FORBIDDEN }
        message: { const: 이 작업을 수행할 권한이 없습니다. }
        request_id: { type: string, format: uuid }
        retryable: { const: false }
    AdminNotFoundError:
      type: object
      additionalProperties: false
      required: [code, message, request_id, retryable]
      properties:
        code: { const: ADMIN_NOT_FOUND }
        message: { const: 대상을 찾을 수 없습니다. }
        request_id: { type: string, format: uuid }
        retryable: { const: false }
    AdminInvalidStateError:
      type: object
      additionalProperties: false
      required: [code, message, request_id, retryable]
      properties:
        code: { const: ADMIN_INVALID_STATE }
        message: { const: 현재 상태에서는 이 작업을 수행할 수 없습니다. }
        request_id: { type: string, format: uuid }
        retryable: { const: false }
    AdminValidationFailedError:
      type: object
      additionalProperties: false
      required: [code, message, request_id, retryable]
      properties:
        code: { const: ADMIN_VALIDATION_FAILED }
        message: { const: 입력값을 확인해 주세요. }
        request_id: { type: string, format: uuid }
        retryable: { const: false }
...
    FailedQuestion:
      type: object
      additionalProperties: false
      x-sejong-state-invariant: failed-question-state-v1
      required: [id, masked_question, intent, fallback_reason, candidate_eligible, status, created_at, text_expires_at, text_purged_at]
      properties:
        id: { type: string, format: uuid }
        masked_question:
          type: [string, 'null']
          description: Null only after the 30-day text retention job has purged the field.
        intent: { $ref: '#/components/schemas/SupportedIntent' }
        fallback_reason: { $ref: '#/components/schemas/StoredFailureReason' }
        candidate_eligible: { type: boolean }
        status: { type: string, enum: [NEW, REASON_CONFIRMED] }
        created_at: { type: string, format: date-time }
        text_expires_at:
          type: string
          format: date-time
          description: Expiry of masked_question only; exactly 30 days after creation.
        text_purged_at:
          type: ['string', 'null']
          format: date-time
          description: Actual purge time; null while masked_question is retained.
      allOf:
        - if:
            properties:
              fallback_reason: { const: INSUFFICIENT_GROUNDING }
            required: [fallback_reason]
          then:
            properties:
              candidate_eligible: { const: true }
          else:
            properties:
              candidate_eligible: { const: false }
        - if:
            properties:
              masked_question: { type: 'null' }
            required: [masked_question]
          then:
            properties:
              text_purged_at: { type: string, format: date-time }
          else:
            properties:
              text_purged_at: { type: 'null' }
    FailedQuestionListResponse:
      type: object
      additionalProperties: false
      required: [items, total]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/FailedQuestion' }
        total: { type: integer, minimum: 0 }
    FailedQuestionDetailResponse:
      type: object
      additionalProperties: false
      required: [item]
      properties:
        item: { $ref: '#/components/schemas/FailedQuestion' }
    CivicScopeGapStatus:
      type: string
      enum: [NEW, PLANNED, DISMISSED]
    CivicScopeGapSummary:
      type: object
      additionalProperties: false
      x-sejong-state-invariant: civic-scope-gap-state-v1
      required: [id, masked_question, status, created_at, updated_at, text_expires_at, text_purged_at, reviewed_by, reviewed_at, review_comment]
      properties:
        id: { type: string, format: uuid }
        masked_question:
          type: [string, 'null']
          minLength: 1
          maxLength: 2000
          description: PII-safe text only; null after exact 30-day retention purge.
        status: { $ref: '#/components/schemas/CivicScopeGapStatus' }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        text_expires_at:
          type: string
          format: date-time
          description: Exactly 30 days after creation.
        text_purged_at: { type: ['string', 'null'], format: date-time }
        reviewed_by: { type: [string, 'null'], minLength: 1, maxLength: 200 }
        reviewed_at: { type: ['string', 'null'], format: date-time }
        review_comment: { type: [string, 'null'], minLength: 1, maxLength: 1000 }
      allOf:
        - if:
            properties:
              status: { const: NEW }
            required: [status]
          then:
            properties:
              reviewed_by: { type: 'null' }
              reviewed_at: { type: 'null' }
              review_comment: { type: 'null' }
          else:
            properties:
              reviewed_by: { type: string, minLength: 1, maxLength: 200 }
              reviewed_at: { type: string, format: date-time }
              review_comment: { type: string, minLength: 1, maxLength: 1000 }
        - if:
            properties:
              masked_question: { type: 'null' }
            required: [masked_question]
          then:
            properties:
              text_purged_at: { type: string, format: date-time }
          else:
            properties:
              text_purged_at: { type: 'null' }
    CivicScopeGapListResponse:
      type: object
      additionalProperties: false
      required: [items, total]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/CivicScopeGapSummary' }
        total: { type: integer, minimum: 0 }
    CivicScopeGapReviewRequest:
      type: object
      additionalProperties: false
      required: [decision, review_comment]
      properties:
        decision: { type: string, enum: [PLANNED, DISMISSED] }
        review_comment: { type: string, minLength: 1, maxLength: 1000 }
    CivicScopeGapReviewResponse:
      type: object
      additionalProperties: false
      required: [id, status]
      properties:
        id: { type: string, format: uuid }
        status: { type: string, enum: [PLANNED, DISMISSED] }
    ReasonConfirmationResponse:
      type: object
      additionalProperties: false
      required: [id, status]
      properties:
        id: { type: string, format: uuid }
        status: { const: REASON_CONFIRMED }
    ReasonConfirmationRequest:
      type: object
      additionalProperties: false
      required: [reason]
      properties:
        reason: { $ref: '#/components/schemas/StoredFailureReason' }
    KBCandidateStatus:
      type: string
      enum: [DRAFTED, PENDING_APPROVAL, APPROVED, REJECTED]
    KBCandidateSummary:
      type: object
      additionalProperties: false
      x-sejong-state-invariant: kb-candidate-state-v1
      required: [id, failed_question_id, title, representative_question, data_origin, category, answer_summary, procedure_steps, required_documents, processing_time, fee, department, source_title, source_url, last_verified_at, caution, status, created_by, reviewed_by, review_comment, approved_at, activated_kb_id, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        failed_question_id: { type: string, format: uuid }
        title: { type: string, minLength: 1, maxLength: 200 }
        representative_question: { type: string, minLength: 1, maxLength: 1000 }
        data_origin: { type: string, enum: [OFFICIAL, MOCK] }
        category:
          type: string
          enum: [MOVE_IN_RESIDENT_REGISTRATION, CERTIFICATE_ISSUANCE, BULKY_WASTE, LOCAL_TAX_GENERAL]
        answer_summary: { type: string, minLength: 1 }
        procedure_steps: { type: array, items: { type: string } }
        required_documents: { type: array, items: { type: string } }
        processing_time: { type: [string, 'null'] }
        fee: { type: [string, 'null'] }
        department: { type: string, minLength: 1 }
        source_title: { type: string, minLength: 1 }
        source_url: { type: string, format: uri, pattern: '^https://' }
        last_verified_at: { type: string, format: date }
        caution: { type: [string, 'null'] }
        status: { $ref: '#/components/schemas/KBCandidateStatus' }
        created_by: { type: string, minLength: 1, maxLength: 100 }
        reviewed_by: { type: [string, 'null'], minLength: 1, maxLength: 100 }
        review_comment: { type: [string, 'null'], minLength: 1, maxLength: 1000 }
        approved_at: { type: [string, 'null'], format: date-time }
        activated_kb_id: { type: [string, 'null'], format: uuid }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
      allOf:
        - if:
            properties:
              status: { enum: [DRAFTED, PENDING_APPROVAL] }
            required: [status]
          then:
            properties:
              reviewed_by: { type: 'null' }
              review_comment: { type: 'null' }
              approved_at: { type: 'null' }
              activated_kb_id: { type: 'null' }
        - if:
            properties:
              status: { const: APPROVED }
            required: [status]
          then:
            properties:
              reviewed_by: { type: string, minLength: 1, maxLength: 100 }
              review_comment: { type: string, minLength: 1, maxLength: 1000 }
              approved_at: { type: string, format: date-time }
              activated_kb_id: { type: string, format: uuid }
        - if:
            properties:
              status: { const: REJECTED }
            required: [status]
          then:
            properties:
              reviewed_by: { type: string, minLength: 1, maxLength: 100 }
              review_comment: { type: string, minLength: 1, maxLength: 1000 }
              approved_at: { type: 'null' }
              activated_kb_id: { type: 'null' }
    KBCandidateListResponse:
      type: object
      additionalProperties: false
      required: [items, total]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/KBCandidateSummary' }
        total: { type: integer, minimum: 0 }
    KBCandidateCreateResponse:
      type: object
      additionalProperties: false
      required: [id, status]
      properties:
        id: { type: string, format: uuid }
        status: { const: DRAFTED }
    KBCandidateSubmitResponse:
      type: object
      additionalProperties: false
      required: [id, status]
      properties:
        id: { type: string, format: uuid }
        status: { const: PENDING_APPROVAL }
    KBCandidateReviewResponse:
      type: object
      additionalProperties: false
      required: [id, status]
      properties:
        id: { type: string, format: uuid }
        status: { type: string, enum: [APPROVED, REJECTED] }
    CandidateReviewRequest:
      type: object
      additionalProperties: false
      required: [decision, review_comment]
      properties:
        decision: { type: string, enum: [APPROVED, REJECTED] }
        review_comment: { type: string, minLength: 1, maxLength: 1000, pattern: '\\S' }
    KBCandidateCreate:
      type: object
      additionalProperties: false
      required: [failed_question_id, title, representative_question, category, answer_summary, department, source_title, source_url, last_verified_at]
      properties:
        failed_question_id: { type: string, format: uuid }
        title: { type: string, minLength: 1, maxLength: 200 }
        representative_question:
          type: string
          minLength: 1
          maxLength: 1000
          description: Human-generalized question that must pass PII validation; not a long-term copy of masked_question.
        category: { $ref: '#/components/schemas/SupportedIntent' }
        answer_summary: { type: string, minLength: 1 }
        procedure_steps: { type: array, items: { type: string } }
        required_documents: { type: array, items: { type: string } }
        processing_time: { type: [string, 'null'] }
        fee: { type: [string, 'null'] }
        department: { type: string, minLength: 1 }
        source_title: { type: string, minLength: 1 }
        source_url: { type: string, format: uri, pattern: '^https://' }
        last_verified_at: { type: string, format: date }
        caution: { type: [string, 'null'] }